fix: TS parity gaps in _parse_session_info_from_lite
Four fixes from a 43-agent adversarial review against the TS reference
(claude-cli-internal src/utils/listSessionsImpl.ts @ main):

1. Tag extraction scoped to {"type":"tag"} lines (sessions.py:455)

   Bare _extract_last_json_string_field(tail, "tag") matches ANY "tag":"..."
   in the 64KB tail — including tool_use inputs from Docker builds, git tag
   via MCP, cloud resource tagging. Proven: test_tag_none_when_only_tool_use_tag
   returned tag='prod' from tool input before the fix.

   TS scopes to lines starting with '{"type":"tag"' at column 0
   (listSessionsImpl.ts:132, sessionStorage.ts:629 with the canonical
   comment about this exact collision class).

   Tests now use separators=(",", ":") to match real on-disk format
   (CLI jsonStringify, SDK tag_session, TS sessionMutationsImpl all write
   compact JSON). 2 regression tests added.

2. custom_title chain: + head scan, + aiTitle fallback (sessions.py:426)

   TS chain is tail.customTitle || head.customTitle || tail.aiTitle ||
   head.aiTitle (listSessionsImpl.ts:97-102). Pre-existing drift from TS
   PRs #21333 (aiTitle) and #20390 landing after Python #622 — fixing here
   since this PR touches the exact block.

3. summary chain: + lastPrompt fallback (sessions.py:436)

   TS inserts tail.lastPrompt between custom_title and summary
   (listSessionsImpl.ts:115-119). Same drift class as (2).

4. created_at: float -> int (types.py:998, sessions.py:465)

   last_modified is int; both are epoch-ms. Wrapped
   datetime.timestamp() * 1000 in int().

+++ src/claude_agent_sdk/types.py
@@ -995,7 +995,7 @@ class SDKSessionInfo:
     git_branch: str | None = None
     cwd: str | None = None
     tag: str | None = None
-    created_at: float | None = None
+    created_at: int | None = None
 
 
 @dataclass
